requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
decorator==5.2.1
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
executing==2.2.0
Flask==3.1.0
fr_core_news_md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.8.0/fr_core_news_md-3.8.0-py3-none-any.whl#sha256=8a70d090a54ef77525c3ffa6a6195b9d365f2cf369ae1cd84ede93f3d709079e
idna==3.10
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.4
packaging==24.2
parso==0.8.4
pdfminer.six==20250327
pexpect==4.9.0
preshed==3.0.9
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
python-docx==1.1.2
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
setuptools==78.1.0
shellingham==1.5.4
smart-open==7.1.0
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
thinc==8.3.6
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
